firmware: Check in initial 'leapmic' firmware
This is a working prototype of the main firmware product. It filters
down the PDM stream from microphones to a nice and cosy 48 ksmps
floating-point PCM stream.

At the moment it's not particularly efficient in its implementation
and it only grabs a single hard-coded microphone channel, but it's an
initial functioning demonstrator.

Signed-off-by: Martin Povišer <povik@cutebit.org>